+Aktualizacja metod w Order oraz dodanie BigDecimal

+++ Task1/src/Product.java
@@ -1,10 +1,11 @@
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Product {
     private int id;
     private String name;
-    private double price;
+    private BigDecimal price;
     private int quantityAvaliable;
     private ProductType type;
     private List<Configuration> configurations = new ArrayList<>();
@@ -12,7 +13,7 @@ public class Product {
     public Product(int id, String name, double price, int quantityAvaliable, ProductType type) {
         this.id = id;
         this.name = name;
-        this.price = price;
+        this.price = BigDecimal.valueOf(price);
         this.quantityAvaliable = quantityAvaliable;
         this.type = type;
     }
@@ -51,11 +52,11 @@ public class Product {
         this.name = name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
